+++ lib/flock.dart
@@ -1,8 +1,11 @@
 library flock;
 
+export 'package:flock/base/store_proxy_base.dart';
+
 export 'base/create_store.dart';
 export 'base/types.dart';
 export 'enhancers/async_publish.dart';
-export 'enhancers/store_proxy_base.dart';
+export 'enhancers/batch_subscribe.dart';
+export 'enhancers/print_events.dart';
 export 'integration/project_to_stream.dart';
 export 'integration/store_builder.dart';